Use a bootstrap and define a trait for shared helpers
