+++ tests/integration/_bootstrap.php
@@ -1,0 +1,3 @@
+<?php
+
+include __DIR__ . '/_helpers/Framework_Integration_Tests_Helper.php';
